+++ src/ExceptionHandlerHelper.php
@@ -166,6 +166,18 @@ class ExceptionHandlerHelper
 		// let's build error message
 		$error_message = '';
 		$ex_message = trim($exception->getMessage());
+
+		// ensure we won't fail due to exception incorect encoding
+		if (!mb_check_encoding($ex_message, 'UTF-8')) {
+			// let's check there's iconv and mb_string available
+			if (function_exists('iconv') && function_exists('mb_detec_encoding')) {
+				$ex_message = iconv(mb_detect_encoding($ex_message, mb_detect_order(), true), 'UTF-8', $ex_message);
+			} else {
+				// lame fallback, in case there's no iconv/mb_string installed
+				$ex_message = htmlspecialchars_decode(htmlspecialchars($ex_message, ENT_SUBSTITUTE, 'UTF-8'));
+			}
+		}
+
 		if (Config::get('response_builder.exception_handler.use_exception_message_first', true)) {
 			if ($ex_message === '') {
 				$ex_message = get_class($exception);