+++ CHANGES.md
@@ -6,6 +6,7 @@ See [compatibility docs](docs/compatibility.md) for details about backward compa
 
 * @dev
    * TestingHelpers trait's `validateResponseStructure()` method is now public
+   * [RB-65] Exception Handler Helper now deals with messages using non-UTF8 or broken encoding 
 
 * v4.0.0 (2017-04-10)
    * **BACKWARD INCOMPATIBILE CHANGES**